:recycle: Refactored gallery and viewer
- Renamed to gallery and viewer for consistency
- Merge two viewer ts files into one
- Tried getting "preview mode" working, but failed
