+++ src/viewer.ts
@@ -1,23 +1,32 @@
 import * as vscode from 'vscode';
 import * as utils from './utils';
 
-export function createPanel(
-	context: vscode.ExtensionContext,
-	imgSrc: string,
-) {
-	const panel = vscode.window.createWebviewPanel(
-		'gryc.viewer',
-		`${utils.getFilename(imgSrc)}`,
-		vscode.ViewColumn.Two,
-		{
+export class ViewerCustomEditor implements vscode.CustomReadonlyEditorProvider {
+	public static readonly viewType = 'gryc.editor';
+
+	constructor(private readonly context: vscode.ExtensionContext) { }
+
+	public async openCustomDocument(uri: vscode.Uri) {
+		return { uri, dispose: () => { } };
+	}
+
+	public async resolveCustomEditor(
+		document: vscode.CustomDocument,
+		webviewPanel: vscode.WebviewPanel,
+	): Promise<void> {
+		let documentPath = webviewPanel.webview.asWebviewUri(document.uri).toString();
+		let documentDir = document.uri.path.split('/').slice(0, -1).join('/');
+		webviewPanel.webview.options = {
+			localResourceRoots: [
+				vscode.Uri.joinPath(this.context.extensionUri, 'media'),
+				vscode.Uri.file(utils.getCwd()),
+                vscode.Uri.file(documentDir),
+            ],
 			enableScripts: true,
-			retainContextWhenHidden: true,
-		}
-	);
-
-	panel.webview.html = getWebviewContent(context, panel.webview, imgSrc);
-
-	return panel;
+			enableForms: false
+		};
+		webviewPanel.webview.html = getWebviewContent(this.context, webviewPanel.webview, documentPath);
+	}
 }
 
 export function getWebviewContent(
